move feature badge
