+++ static/app/utils/metrics/features.tsx
@@ -1,6 +1,4 @@
-import FeatureBadge from 'sentry/components/badge/featureBadge';
 import type {Organization} from 'sentry/types/organization';
-import useOrganization from 'sentry/utils/useOrganization';
 import {Dataset} from 'sentry/views/alerts/rules/metric/types';
 
 export function hasMetricsExperimentalFeature(organization: Organization) {
@@ -24,16 +22,6 @@ export function hasRolledOutMetrics(organization: Organization) {
 
 export function canSeeMetricsPage(organization: Organization) {
   return hasCustomMetrics(organization) || hasRolledOutMetrics(organization);
-}
-
-export function MetricsFeatureBadge() {
-  const organization = useOrganization();
-
-  if (hasRolledOutMetrics(organization)) {
-    return <FeatureBadge type="new" />;
-  }
-
-  return <FeatureBadge type="beta" />;
 }
 
 /**
